+++ finetrainers/ltx_video/__init__.py
@@ -1,1 +1,2 @@
-from .ltx_video_lora import LTX_VIDEO_T2V_FULL_FINETUNE_CONFIG, LTX_VIDEO_T2V_LORA_CONFIG
+from .full_finetune import LTX_VIDEO_T2V_FULL_FINETUNE_CONFIG
+from .ltx_video_lora import LTX_VIDEO_T2V_LORA_CONFIG
